% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlva.R
\name{extract.amplicon.multiple.sequence}
\alias{extract.amplicon.multiple.sequence}
\title{extract.amplicon.multiple.sequence enables to extract from a target sequence (a DNAstringset of lenght n) the amplicon(s) defined by the primer pair}
\usage{
extract.amplicon.multiple.sequence(target, Fprobe, Rprobe, max.mismatch)
}
\arguments{
\item{target}{the reference sequence that you want to screen}

\item{Fprobe}{the sequence of the forward probe}

\item{Rprobe}{the sequence of the reverse probe}

\item{max.mismatch}{the maximum number of mismatches allowed in the probes}
}
\value{
the sequence of the amplicon (including both primers)
}
\description{
You provide both a genome and the probes
}
