% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlva.R
\name{type1motif}
\alias{type1motif}
\title{type1motif enables to compute the number of motif for a single VNTR locus}
\usage{
type1motif(genome, motif, kmer)
}
\arguments{
\item{genome}{the reference genome that you want to screen}

\item{kmer}{the kmer size which was used during genome assembly}

\item{motifs}{the sequence of the motif}
}
\value{
the number of repetitions
}
\description{
You provide both a genome and the motif. The function computes the number of repetition for this motif
}
