% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlva.R
\name{typemlva}
\alias{typemlva}
\title{typemlva enables to compute the mlva profile of an isolate based on the genome}
\usage{
typemlva(genome, motifs, kmer)
}
\arguments{
\item{genome}{the reference genome that you want to screen}

\item{motifs}{a list with each motif}

\item{kmer}{the kmer size which was used during genome assembly}
}
\value{
a characther with the MLVA profile
}
\description{
You provide both a genome and the motifs. The function computes the number of repetition for each motif
}
